services/api_gateway/requirements.txt
# 核心依赖
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
prometheus-client==0.16.0

# HTTP客户端
httpx==0.24.0

# 认证和安全
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# 缓存
redis==4.5.5

# 日志和监控
python-json-logger==2.0.7

# 数据库
motor>=2.5.0

# 监控和追踪
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-instrumentation-fastapi==0.30b1
opentelemetry-exporter-jaeger==1.11.1

# 工具
python-consul>=1.1.0
python-dotenv>=0.19.0
loguru>=0.5.3 